Move EvaluationResult type into teaching lesson model
